SystemIcons Example now really showing the system icons, not only the first in kListboxContent
I realized SystemIcons example was only showing QuicklookTemplate, the
first of the icons listed in kListboxContent.
I removed the residing size informations (like :16:12) in a few of the
constant’s lines and added another loop to LB1’ open method, but let
the first one (although it is only run through once) in case there is
some deeper meaning to it.